requirements.txt
autopep8==2.0.1
charset-normalizer==3.0.1
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
distlib==0.3.6
docopt==0.6.2
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.31
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.0
mplcursors==0.5.2
mpld3==0.5.9
numpy==1.24.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pycodestyle==2.10.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
scikit-learn==1.2.1
scipy==1.10.1
six==1.16.0
sklearn==0.0.post1
smmap==5.0.0
threadpoolctl==3.1.0
tomli==2.0.1
tqdm==4.64.1
virtualenv-clone==0.5.7
